settings: Remove is_nested parameter and disableable class.
The is_nested paramter in settings_checkbox.hbs used to
set disableable class in the div element is not used
anywhere and thus we can remove both the is_nested parameter
and disableable class.

The is_nested paramater was first added in ad2effe6c6 and
disableable class was added in 997fe77b5.

The use of is_nested parameter was removed in 8d02cd93a2789.