notifications: Don't enqueue notifications for bots.
This replaces the temporary (and testless) fix in
c9c7594d81eb2eb96c19fdc96b6b3e04afbeee53 with a more permanent
fix.

Instead of checking if the user is a bot just before
sending the notifications, we now just don't enqueue
notifications for bots. This is done by sending a list
of bot IDs to the event_queue code, just like other
lists which are used for creating NotificationData objects.

Credit @andersk for the test code in `test_notification_data.py`.