compose: Fix compose box initialization order.
This effectively reverts part of
feae715438a86a6de3a7d8c1c4a5323cbab8c579.  While it's correct that we
want to render this bit of Handlebars template early, it was not
correct to move all compose box initialization earlier.

Do the same thing we do with the left/right sidebar container
templates, which is to render them directly in `ui_init.js`.

Fixes #20778.