bots: Do not allow to remove owner from edit bot form in personal settings.
bb172d8c02d removed the option to remove owner from the edit form available
to admins, but the option was still present in the edit form available to
the owner of bots in "Bots" section of personal settings. This commit
removes the option from form in "Bots" section of personal settings as well.