puppeteer: Resurrect test_change_password.
c4c1227 removed this test as it was causing alert words test
fail non-deterministically. See c4c1227's commit message for
some more details.

Rearranging the order to make this test run in the end can
eliminate the non-deterministic failures due to this test.

Because we reset the test database in between test files, we don't
expect there to be any issues with test files running after this.