streams: Mark messages as read only for unsubscribed streams.
Previously, we were marking messages of all the streams passed
to bulk_remove_subscriptions even if user was not subscribed
to some of them and those streams would ideally not have
any unread messages. This code was added in 2fbad1c9af8.

This commit changes the code to only mark messages of actually
unsubscribed streams as read.