app-loading: Add a delay before showing the reload message.
Fixes #22182

This message often flashes on screen briefly, causing unnecessary
worry for the user (is the app likely to not load?).
To address this, we add a delay before the message is shown.

As a consequence, we change the notice to no longer suggest waiting a
few seconds, since we did that before showing it.