stream_settings: Modify the message of the "Announce stream" hint tooltip.
Follow-up to commit 3d25d7c03b4b07a948d19af688cddd95f990278e.

The "Announce stream" hint tooltip earlier read "Organization
administrators can change this in in the organization settings."
It wasn't obvious that "this" refers to the stream the notification
will go to, so the tootltip message has been modified to reflect this.