check-rabbitmq-consumers: Remove unused --min-threshold.
This has never actually been used  -- and does not make sense with the
check-all-queues-at-once model switched to in ae5274a43d90.  The
Tornado processes are the only ones we expect to be non-1, and since
they were added in 8a3aa58153ed the right number has been read from
config, not passed as an argument.