help-docs: Document "Resolve/Unresolve topic" mobile app feature.
Adds a new tab with the resolve/unresolve topic instructions for
mobile app users.

Makes the instructions a little more explicit so that users know they
won't see the menu unless they press and hold the topic long enough.

Also improves the wording of the instructions to access the
long-press menu so that users are more likely to read it as saying
that the whole area of a given topic, extending the whole width of
the screen, is the region they can press to act on that topic.

Fixes #22144.